Better thread view

+++ GameEngine/GameEngine/Engine/Utils/Age_Imgui.cpp
@@ -33,7 +33,7 @@ namespace AGE
 		ImGuiIO& io = ImGui::GetIO();
 		//HARDCODED WINDOW TO FIX
 		auto screenSize = en->getInstance<IRenderContext>()->getScreenSize();
-		io.DisplaySize = ImVec2(screenSize.x, screenSize.y);        // Display size, in pixels. For clamping windows positions.
+		io.DisplaySize = ImVec2((float)screenSize.x, (float)screenSize.y);        // Display size, in pixels. For clamping windows positions.
 		io.DeltaTime = 1.0f / 60.0f;                          // Time elapsed since last frame, in seconds (in this sample app we'll override this every frame because our timestep is variable)
 		io.PixelCenterOffset = 0.0f;                        // Align OpenGL texels
 		io.KeyMap[ImGuiKey_Tab] = SDL_SCANCODE_TAB;             // Keyboard mapping. ImGui will use those indices to peek into the io.KeyDown[] array.
